StyleProvider: Convert component to TypeScript
