+++ packages/components/src/style-provider/types.ts
@@ -1,0 +1,15 @@
+/**
+ * External dependencies
+ */
+import type { ReactNode } from 'react';
+
+export type StyleProviderProps = {
+	/**
+	 * The children elements.
+	 */
+	children: ReactNode;
+	/**
+	 * Current document.
+	 */
+	document: Document;
+};
